+++ integration-tests/src/lib.rs
@@ -4,7 +4,7 @@ pub mod contract;
 pub mod coordinator_contract;
 pub mod gateway_contract;
 pub mod multisig_contract;
-pub mod multisig_prover_contract;
+pub mod solana_multisig_prover_contract;
 pub mod protocol;
 pub mod rewards_contract;
 pub mod router_contract;